pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.18.0",
    "google-auth>=2.43.0",
    "google-cloud-storage>=3.6.0",
    "openpyxl>=3.1.5",
    "pillow>=12.0.0",
]
